Slate 2.0 - Break up engine
