+++ wear/src/main/java/ca/joshstagg/slate/complication/RenderExtension.kt
@@ -17,28 +17,28 @@ internal fun Render.renderText(context: Context, icon: Icon?) {
 
     when {
         icon != null -> {
-            val rect = rect
-            val drawable = icon.loadDrawable(context)
-            var widthOffset = rect.width() / 3
-            var heightOffset = rect.height() / 3
-            var top = heightOffset
-            var bottom = heightOffset
+            icon.loadDrawable(context)?.let { drawable->
+                var widthOffset = rect.width() / 3
+                var heightOffset = rect.height() / 3
+                var top = heightOffset
+                var bottom = heightOffset
 
-            mainText?.let {
-                widthOffset = rect.width() * 3 / 8
-                heightOffset = rect.height() * 1 / 8
-                top = 2 * heightOffset
-                bottom = 4 * heightOffset
-                y += 2 * heightOffset
+                mainText?.let {
+                    widthOffset = rect.width() * 3 / 8
+                    heightOffset = rect.height() * 1 / 8
+                    top = 2 * heightOffset
+                    bottom = 4 * heightOffset
+                    y += 2 * heightOffset
+                }
+
+                drawable.setTint(paints.complicationTint)
+                drawable.setBounds(
+                        rect.left + widthOffset,
+                        rect.top + top,
+                        rect.right - widthOffset,
+                        rect.bottom - bottom)
+                drawable.draw(canvas)
             }
-
-            drawable.setTint(paints.complicationTint)
-            drawable.setBounds(
-                    rect.left + widthOffset,
-                    rect.top + top,
-                    rect.right - widthOffset,
-                    rect.bottom - bottom)
-            drawable.draw(canvas)
         }
         subText != null -> {
             val subMessage = subText.getText(context, currentTimeMills)
