+++ wear/src/main/res/layout/config_checkbox_row.xml
@@ -4,11 +4,10 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:foreground="?android:attr/selectableItemBackground"
+    android:paddingBottom="8dp"
     android:paddingEnd="@dimen/config_activity_padding"
     android:paddingStart="@dimen/config_activity_padding"
-    android:paddingTop="8dp"
-    android:paddingBottom="8dp"
-    >
+    android:paddingTop="8dp">
 
     <CheckBox
         android:id="@+id/config_checkbox"
@@ -16,9 +15,9 @@
         android:layout_width="40dp"
         android:layout_height="40dp"
         android:layout_centerVertical="true"
-        android:gravity="center"
         android:background="@drawable/ws_switch_track_mtrl"
-        android:backgroundTint="@color/ws_switch_track_color_material"/>
+        android:backgroundTint="@color/ws_switch_track_color_material"
+        android:gravity="center" />
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -49,6 +48,7 @@
             android:ellipsize="end"
             android:gravity="center_vertical|start"
             android:maxLines="1"
+            android:textColor="@color/secondary_text_light"
             android:textSize="@dimen/summary_text_size" />
 
     </LinearLayout>
